requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
lxml==4.5.2
networkx==2.5
python-libsbml==5.18.0
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
